Fee-Collection-Software_Balkanjibari_Nadiad-2026-main/backend/requirements.txt
APScheduler==3.11.2
argon2-cffi==25.1.0
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-csp==4.0
django-decouple==2.1
django-filter==25.2
django-formtools==2.5.1
django-otp==1.7.0
django-otp-yubikey==1.1.0
django-phonenumber-field==8.4.0
django-ratelimit==4.1.0
django-two-factor-auth==1.18.1
pyotp==2.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-database-url==3.1.0
drf-spectacular==0.29.0
gunicorn==25.1.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.2
pillow==12.1.0
pandas==3.0.2
numpy==2.4.4
openpyxl==3.1.5
qrcode==8.2
python-docx==1.2.0
python-pptx==1.0.2
PyYAML==6.0.3
requests==2.32.5
whitenoise==6.11.0
razorpay==2.0.0
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
tzdata==2025.3
tzlocal==5.3.1
phonenumbers==8.13.55
pycryptodome==3.23.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
certifi==2026.1.4
idna==3.11
charset-normalizer==3.4.4
typing_extensions==4.15.0
setuptools==80.10.2
